+++ android/app/src/main/java/org/uwaterloo/subletr/api/models/ListingSummary.kt
@@ -36,7 +36,7 @@ import com.squareup.moshi.JsonClass
 data class ListingSummary (
 
     @Json(name = "listing_id")
-    val listingId: kotlin.Long,
+    val listingId: kotlin.Int,
 
     @Json(name = "address")
     val address: kotlin.String,
